A pile of updates.
Many of these updates are related to gnus.  I have a love-hate
relationship with this thing.  I used it for a long time, but damnit,
it just doesn't quite satisfy me.  I yearn for a mail and newsreader
for Mac OS X that's as powerful as gnus, but not as junky and slow and
quirky.  *sigh*

See also

  http://emacs-fan.blogspot.com/2011/03/paredit-keeping-your-lisp-code-proper.html

for my adventures with paredit.